Prefer online players when using /jailinfo
Previously /jailinfo player would always get the player's info straight
from the database, even if there was somebody online with the name,
which would make it impossible to get /jailinfo on a player if player
had changed their name, due to the fact that the playername in the Jail
database records the name the player had when jailed, and is not
updated.

I.e. 'player' is jailed, and an entry is added to the jail database with
the name 'player'. If 'player' then changes name to 'changed', then
previously /jailinfo changed would try to search the database for
anybody named 'changed', returning no results. Now it will see that a
player named changed is online, get their UUID, and search through the
jail database with the UUID instead of the name.